updated the code

+++ disks.hpp
@@ -168,7 +168,7 @@ sorted_disks sort_lawnmower(const disk_state &before) {
   int numOfSwap = 0;
   disk_state step = before;
     
-    for(int i = 0; i < step.light_count(); i++) {
+    for(size_t i = 0; i < step.light_count(); i++) {
     // left to right - compares every two adjacent disks and swaps if necessary
     for (size_t j = 0; j < step.total_count() - 1; j++) {
       if (step.get(j) > step.get(j + 1)) {
